+++ test/services/schema-file-updater.test.js
@@ -23,7 +23,7 @@ describe('services > schema-file-updater', () => {
   const buildSchema = (collection, metas = meta) =>
     new SchemaFileUpdater('test.json', collection, metas, serializerOptions).perform();
 
-  // NOTICE: Expecting fs.writeFileSync second parameter to be valid JSON
+  // NOTICE: Expecting `fs.writeFileSync` second parameter to be valid JSON.
   it('should call fs.writeFileSync with a valid JSON as data', () => {
     expect.assertions(2);
     buildSchema([], {});